[r3] Re-design choices, bounding boxes, etc...
Re-designed the whole structure of the program so, that header files
only and only will be used as headers that define the structure of a
class (Math classes are exceptions) and the actual function declarations
will be in separate .cpp source files, as it should be. This is not an
optimization related danger anymore because of LTO that can be enabled
for the compiler and linker, which will allow inlining of functions
between different .o files.

Also, created a few more core classes such as AABB (Axis aligned
bounding box), etc...